requirements.txt
# FastAPI framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlmodel>=0.0.22
asyncpg>=0.30.0
greenlet>=3.1.0

# Validation and settings
pydantic>=2.10.0
pydantic-settings>=2.6.0

# JWT Authentication
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0

# Code quality
black>=24.10.0
ruff>=0.8.0
